captureVideo.json config files added...
Added in (3) files, two of which are for reference and archival.

The main file is "captureVideo.json".  This needs to be in the same directory as the "captureVideo.py" script.  There are a few more vars I should probably pull into the config (like the full path to the video capture dump directory), but for the moment it appears to be fully functional...