Updating config files and main script...
Added "OutputExtension" to config files and added zero-mod script support to captureVideo.py.

captureVideo.py now auto-detects the model name of the machine that it's running on and looks for a corresponding config file.  This allows a user to change the video recording parameters without having to modify or customize the main script.  

The process for bringing on a new machine type will be to manually "tune" the video capture parameters, create a corresponding config file (based on the data in "/proc/device-tree/model"), and drop this file in the same directory as captureVideo.py.