Updating 'capture.py'...
Added in a "capture.stop" flag file detection.  If any file named "capture.stop" (case sensitive) is detected in the same directory as the "capture.py" script, the script will stop after it's latest recording.

Added a "SpaceSaver" function to compress files after they've been created and released.  This is hacked-in solution with a lot of problems for in some scenarios, but hopefully it will function as needed for the first deployment.

SpaceSaver **MUST BE MOVED** to a separate service so the processing rate can be throttled!  Currently, as soon as a file is completed and released, it spawns a SpaceSaver call.  However, if the SpaceSaver routine takes longer than the average clip time to record, you can end up with multiple stacked instances of FFMPEG runnig in the background.  This is likely to cause the system to crash and restart.

WE MAY NEED TO DISABLE SPACESAVER FOR THE INITIAL BUOY DEPLOYMENT FOR THIS REASON!!!

Further testing is required.  

Naturally...