Add files via upload
Mostly updates on installer stuff, but also the beginnings of a few other things.

WARNING -  This is currently a disorganized mess and there is a TON of crap that needs to be cleared out of here.  That's to say nothing of the horridly sloppy directory structuring I've drooled all over this thing like a slack-jawed ding-dong.

Buuuuuutttttt....that's the price of R&D!  Right?  Right...